+++ index.html
@@ -7,7 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="js/functions.js" defer></script>
-  <script src="js/main.js" defer></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -236,5 +235,6 @@
       </section>
     </template>
 
+    <script src="js/main.js" type="module"></script>
 </body>
 </html>
