+++ package.json
@@ -2,6 +2,7 @@
   "name": "kekstagram",
   "version": "30.0.0",
   "private": true,
+  "type": "module",
   "description": "Личный проект «Кекстаграм» от HTML Academy",
   "repository": {
     "type": "git",
